modify folder structure and ensure parent deck limits are sum of child decks
