+++ managers/__init__.py
@@ -1,0 +1,10 @@
+# Copyright (C) 2025
+# Managers Package for AnkiConnect
+#
+# This package contains manager classes for handling different
+# aspects of Anki functionality.
+
+from .study_manager import StudyManager
+from .note_manager import NoteManager
+
+__all__ = ['StudyManager', 'NoteManager']
