+++ managers/note_manager.py
