+++ manifest.json
@@ -1,5 +1,5 @@
 {
   "package": "AnkiConnect",
-  "name": "Global Edge Anki Connect v0.0.5",
-  "version": "0.0.5"
+  "name": "Global Edge Anki Connect v0.0.6",
+  "version": "0.0.6"
 }
